bes2600: Update the wifi driver from bestechnic (0.3.5_2023.0209)
Import from https://gitlab.com/TuxThePenguin0/bes2600

New driver intended to work with Linux 6.3 provided by bestechnic.
Despite the lower version number this does appear to be a newer version.

+++ drivers/staging/bes2600/ap.h
@@ -26,7 +26,7 @@ void bes2600_sta_notify(struct ieee80211_hw *dev, struct ieee80211_vif *vif,
 void bes2600_bss_info_changed(struct ieee80211_hw *dev,
 			     struct ieee80211_vif *vif,
 			     struct ieee80211_bss_conf *info,
-			     u32 changed);
+			     u64 changed);
 
 int bes2600_ampdu_action(struct ieee80211_hw *hw,
 			struct ieee80211_vif *vif,
